+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add Tile transformation
   (<https://github.com/openvinotoolkit/datumaro/pull/790>)
+- Add Video keyframe extraction
+  (<https://github.com/openvinotoolkit/datumaro/pull/791>)
 
 ### Changed
 - Improved mask_to_rle performance
@@ -23,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix MacOS CI failures
   (<https://github.com/openvinotoolkit/datumaro/pull/789>)
+- Fix auto-documentation for the data_format plugins
+  (<https://github.com/openvinotoolkit/datumaro/pull/793>)
 - Fix CVAT for images 1.1 data format import at project level
   (<https://github.com/openvinotoolkit/datumaro/pull/795>)